создание тестов opencart.
